v1.5.19 Support fractional scores and improve JSON error handling
Changed PlayerRanking.Score to double? and Ranking properties to double for fractional values. Added try-catch in LoadJsonAsync.cs for better deserialization error reporting.

+++ src/Smab.TTInfo.TTLeagues/Models/TTLeagues/PlayerRanking.cs
@@ -13,7 +13,7 @@
 /// <param name="Category">The category in which the player is ranked. This is a non-null string that identifies the context of the ranking,
 /// such as "Sports", "Gaming", or "Academics".</param>
 public sealed record PlayerRanking(
-	int?   Position,
-	int?   Score,
-	string Category
+	int?    Position,
+	double? Score,
+	string  Category
 );
